+++ activemodel/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Add `ActiveModel::Conversion.param_delimiter` to configure delimiter being used in `to_param`
+
+    *Nikita Vasilevsky*
+
 *   `undefine_attribute_methods` undefines alias attribute methods along with attribute methods.
 
     *Nikita Vasilevsky*